Created a Player, Camera Controller, and Obstacles
Very simple infinite runner where the player runs, infinitely. The basis is in there. The player will infinitely run right (I know this is handled in a different way where the objects are spawned). This was so that momentum could be made. Cool momentum stuff. Then now the player can jump and not infinitely. They can run into objects and slow down.  Camera will zoom in and out based on player speed as well.