Chunk phases and checkpoints completed
Phases can now be properly generated, spawning in specific chunks at specific phases. These are easily adjustable inside of the inspector for the ChunkManager.

Checkpoints are made in a generic CheckpointChunk. Checkpoint script was made to increment a score manager.

Score Manager is made to both display text and keep track of checkpoints.

Hoo baby game loop is basically done